use old rng everywhere
