+++ src/Random.js
@@ -1,7 +1,23 @@
 export class Random
 {
-    constructor() {
-        this.seed = Date.now();
+    constructor(seed) {
+        function getHashCode(value) { // https://stackoverflow.com/a/7616484
+            var hash = 0,
+                i, chr;
+            if (value.length === 0) return hash;
+            for (i = 0; i < value.length; i++) {
+                chr = value.charCodeAt(i);
+                hash = ((hash << 5) - hash) + chr;
+                hash |= 0; // Convert to 32bit integer
+            }
+            return hash;
+        }
+        
+        if (!seed) {
+            this.seed = Date.now();
+        } else {
+            this.seed = getHashCode(seed);
+        }
     }
 
     setSeed(seed) {
@@ -12,8 +28,7 @@ export class Random
         this.seed = Date.now();
     }
 
-    // named Number to match built in boardgame.io random.Number()
-    Number() {
+    next() {
         var x = Math.sin(this.seed++) * 10000;
         return x - Math.floor(x);
     }
